Refactored command handling.
Command are now registered in a factory and added to a menu in one go. A
single slot in MainWindow is to be used to handle called actions and
commands execution.

A new EvaluationContext struct was added to pass around information
relative to the context in which the command are executed (for now only
a pointer to the scene is passed).