Refactor object creation and editing using a node system.
The (basic) modifiers were replaced by a node based system for better
control and flexibility over the creation of objects.

The object data was split in a separate Primitive class. Nodes can
create and modify primitives which are passed around in the object's
graph, the last primitive in the graph (passed to the output node) is
then set as the primitive of the object. The type of the final primitive
will be the type of the object.

The primitives created inside of a graph are reference counted and
deleted when they are not used anymore. Purging of the primitives
happens for now before each time the node graph is evaluated.