Improve node editor view zooming and scrolling.
The QGraphicsView was subclassed to be able to reimplement its
wheelEvent method which effectively disconnects the mouse wheel from the
scroll bars. So there is no more conflicts between zooming and
scrolling.

It is also possible to scroll the view while dragging the cursor inside
of it (through left click).