Refactor the way properties are added to objects and exposed in the UI.
Now properties, for Nodes at the moment, are stored in the object that
holds them, which will allow to update property flags to show/hide props
in the UI.

To add some properties to an object, it must be derived from Watched
class. Which will also be used for general MVC purposes, and also
provides an API to update UI params in a generic way.

This patch also includes a huge cleanup of the node item's code.