Turn object into a container which can hold multiple primitives.
This commit introduces a new way of handling primitives. The primitives
are now held inside a PrimitiveCollection contained inside of an object.

The main reason for this change is to be able to support dynamic objects
like fractured objects or results of simulations like fluids simulations
which may contain multiple volumes (fields).

The majority of the code which used to work on primitives now work on
deals with this PrimitiveCollection class. The collection is also to be
used to create new primitive instances, and simplifies passing data
around in node graphs.

Since multiple primitives can now be inside of the same object, the
Primitive class has been augmented with a type id which needs to be set
by the plugins registering new primitive types, and can be used during
loops over the collection to check for the primitive type.