Complete refactor of the UI widget and update system.
The UI code is now somewhat properly separated from the core code.

This new system consist of various widgets (timeline, outliner, graph
editor, ...) which listen to the context for events. It is now possible
to have multiple instance of the same widget, which are all kept in
sync. Those widgets were split from the mainwindow signal and slots
mechanism, and the UI form is now removed.

The old wathcer/watched classes are replaced with a single class called
Persona will is the base calss to derive objects with UI props from.

There is still a lot to be done in terms of update or redraws,
especially for the graph editor and the outliner, but it is a good start
for a cleaner MVC architecture.