Evaluate object graphs as part of the dependency graph evaluation.
This adds a new node type to the dependency graph (ObjectGraphDepsNode)
which takes care of building and evaluating the object graph. This node
is now the parent of the object dependency node. The old eval_graph
function was split between the object and object graph dependency nodes
process methods.

This also adds a new TimeDepsNode which will in the future be used to
parent nodes which depends on time to, and evaluate nodes which depends
on it when the scene frame changes.