Add a SceneNode base class to derive scene level objects from.
This commit makes the Object derive from a base node class, with inputs
and outputs which are drawn in the graph editor. For now there is only
one input and one output associated with the Object. Connecting these in
the graph editor will parent the objects together, with the "output" one
used as the parent and the "input" one the child.

To make sure Objects can both depend on another one from parenting, and
depend on scene time, depsgraph nodes' input sockets now support more
than one connection.

This commit also brings a few changes to the outliner, namely:
- highlight outliner item for the active node in the scene.
- add object nodes in the outliner.
- flags added to Scene and SceneNode to store state of outliner widget
expansion, used when recreating the items upon updates.

Initial support for parenting objects by drag & drop in the outliner is
present in the patch, but disabled as buggy/unfinished.