Initial support for key maps.
This adds a global variable, hidden inside a namespace, that gathers key
maps. Those key maps contain a set of keys and modifier keys coupled
with the command that they are supposed to invoke.

The goal is to make it so a unified system will call commands based on
mouse and keyboard inputs so that we can have good support for undo and
redo.

Undo and redo will be achieved by means of copying the scene and its
data before the command is called by the global command manager.