Refactor the node editor's draw update logic.
Graphs are now drawn from scratch every time the context is changed.

This will also make it easier to move some logic like connecting nodes
outside of the UI code. Then we will also get rid of the edit_mode flag
and make editors sensitive to the current node in the hierarchy.

The scene and object nodes are now storing the position inside the
editor as well as their selection state.